+++ gaphor/SysML/drop.py
@@ -3,6 +3,8 @@ from gaphor.diagram.presentation import connect
 from gaphor.diagram.support import get_diagram_item
 from gaphor.SysML.sysml import ProxyPort
 from gaphor.UML.drop import diagram_has_presentation
+from gaphor.SysML.blocks.property import PropertyItem
+from gaphor.SysML.sysml import Block
 
 
 @drop.register
@@ -11,7 +13,23 @@ def drop_proxy_port(element: ProxyPort, diagram, x, y):
     if not item_class:
         return None
 
-    head_item = diagram_has_presentation(diagram, element.encapsulatedClassifier)
+    def property_item_which_is_typed_by_block_owning_the_port(item):
+        return (
+            isinstance(item, PropertyItem)
+            and isinstance(item.subject.type, Block)
+            and element in item.subject.type.ownedPort
+        )
+
+    head_item = next(
+        (
+            item
+            for item in diagram.select(
+                property_item_which_is_typed_by_block_owning_the_port
+            )
+        ),
+        diagram_has_presentation(diagram, element.encapsulatedClassifier),
+    )
+
     if not head_item:
         return None